+++ src/ViewInstance.js
@@ -65,6 +65,7 @@ import ViewHoldingsRecord from './ViewHoldingsRecord';
 import ViewMarc from './ViewMarc';
 import makeConnectedInstance from './ConnectedInstance';
 import withLocation from './withLocation';
+import InstancePlugin from './components/InstancePlugin';
 
 import {
   HoldingsListContainer,
@@ -127,6 +128,7 @@ class ViewInstance extends React.Component {
 
     this.state = {
       marcRecord: null,
+      findInstancePluginOpened: false,
       isItemsMovement: false,
     };
     this.instanceId = null;
@@ -185,6 +187,12 @@ class ViewInstance extends React.Component {
       search: location.search,
     });
   };
+
+  selectInstanse = (instance) => {
+    const { location: { search } } = this.props;
+
+    this.props.goTo(`/inventory/view/${instance.id}${search}`);
+  }
 
   toggleItemsMovement = () => {
     this.setState((prevState) => ({ isItemsMovement: !prevState.isItemsMovement }));
@@ -302,6 +310,10 @@ class ViewInstance extends React.Component {
     return ref || {};
   };
 
+  toggleFindInstancePlugin = () => {
+    this.setState(prevState => ({ findInstancePluginOpened: !prevState.findInstancePluginOpened }));
+  }
+
   createActionMenuGetter = instance => ({ onToggle }) => {
     const {
       onCopy,
@@ -397,6 +409,19 @@ class ViewInstance extends React.Component {
             <FormattedMessage
               id={`ui-inventory.moveItems.instance.actionMenu.${this.state.isItemsMovement ? 'disable' : 'enable'}`}
             />
+          </Icon>
+        </Button>
+
+        <Button
+          id="move-instance"
+          buttonStyle="dropdownItem"
+          onClick={() => {
+            onToggle();
+            this.toggleFindInstancePlugin();
+          }}
+        >
+          <Icon icon="arrow-right">
+            <FormattedMessage id="ui-inventory.moveItems" />
           </Icon>
         </Button>
       </Fragment>
@@ -746,7 +771,7 @@ class ViewInstance extends React.Component {
     return (
       <Pane
         data-test-instance-details
-        defaultWidth={paneWidth}
+        style={{ flex: 'auto' }}
         appIcon={<AppIcon app="inventory" iconKey="instance" />}
         paneTitle={
           <span data-test-header-title>
@@ -1415,6 +1440,12 @@ class ViewInstance extends React.Component {
           />
         */ }
         <Callout ref={this.calloutRef} />
+        {this.state.findInstancePluginOpened
+          && <InstancePlugin
+            onSelect={this.selectInstanse}
+            onClose={this.toggleFindInstancePlugin}
+            withTrigger={false}
+          />}
       </Pane>
     );
   }